refactor(cli): container and host vulnerability cmds
